requirements.txt
fastapi
uvicorn
python-multipart
pymupdf
openai
python-dotenv
pandas
openpyxl